+++ rlhfblender/data_models/global_models.py
@@ -1,6 +1,6 @@
 from typing import List
-import numpy as np # for eval() we might have to import numpy
 
+import numpy as np  # for eval() we might have to import numpy
 from pydantic import BaseModel, field_validator
 
 """
