--- rlhfblender/utils/video_callback.py
@@ -1,74 +1,0 @@
-import os
-from typing import Any, Dict
-
-import gymnasium as gym
-import torch as th
-import torchvision
-from stable_baselines3.common.callbacks import BaseCallback
-from stable_baselines3.common.evaluation import evaluate_policy
-from stable_baselines3.common.logger import Video
-
-
-class VideoRecorderCallback(BaseCallback):
-    def __init__(
-        self,
-        eval_env: gym.Env,
-        render_freq: int,
-        n_eval_episodes: int = 2,
-        deterministic: bool = True,
-        is_airl=False,
-    ):
-        """
-        Records a video of an agent's trajectory traversing ``eval_env`` and logs it to TensorBoard
-
-        :param eval_env: A gym environment from which the trajectory is recorded
-        :param render_freq: Render the agent's trajectory every eval_freq call of the callback.
-        :param n_eval_episodes: Number of episodes to render
-        :param deterministic: Whether to use deterministic or stochastic policy
-        """
-        super().__init__()
-        self._eval_env = eval_env
-        self._render_freq = render_freq
-        self._n_eval_episodes = n_eval_episodes
-        self._deterministic = deterministic
-
-    def _on_step(self) -> bool:
-        if self.n_calls % self._render_freq == 0:
-            screens = []
-
-            def grab_screens(_locals: Dict[str, Any], _globals: Dict[str, Any]) -> None:
-                """
-                Renders the environment in its current state, recording the screen in the captured `screens` list
-
-                :param _locals: A dictionary containing all local variables of the callback's scope
-                :param _globals: A dictionary containing all global variables of the callback's scope
-                """
-                screen = self._eval_env.render()
-                # PyTorch uses CxHxW vs HxWxC gym (and tensorflow) image convention
-                screens.append(screen)
-
-            evaluate_policy(
-                self.model,
-                self._eval_env,
-                callback=grab_screens,
-                n_eval_episodes=self._n_eval_episodes,
-                deterministic=self._deterministic,
-            )
-            self.logger.record(
-                "trajectory/video",
-                Video(
-                    th.ByteTensor([[screen.transpose(2, 0, 1) for screen in screens]]),
-                    fps=25,
-                ),
-                exclude=("stdout", "log", "json", "csv"),
-            )
-            print(
-                "save video to: ",
-                os.path.join(self.logger.dir, "video_episode_" + str(self.n_calls)),
-            )
-            torchvision.io.video.write_video(
-                os.path.join(self.logger.dir, "video_episode_" + str(self.n_calls) + ".mp4"),
-                screens,
-                fps=25,
-            )
-        return True
