--- rlhfblender/utils/wrappers.py
@@ -1,311 +1,0 @@
-import gymnasium as gym
-import numpy as np
-
-# (backward compatibility)
-from scipy.signal import iirfilter, sosfilt, zpk2sos
-
-
-class DoneOnSuccessWrapper(gym.Wrapper):
-    """
-    Reset on success and offsets the reward.
-    Useful for GoalEnv.
-    """
-
-    def __init__(self, env: gym.Env, reward_offset: float = 0.0, n_successes: int = 1):
-        super().__init__(env)
-        self.reward_offset = reward_offset
-        self.n_successes = n_successes
-        self.current_successes = 0
-
-    def reset(self):
-        self.current_successes = 0
-        return self.env.reset()
-
-    def step(self, action):
-        observation, reward, terminated, truncated, info = self.env.step(action)
-        if info.get("is_success", False):
-            self.current_successes += 1
-        else:
-            self.current_successes = 0
-        # number of successes in a row
-        done = terminated or truncated or self.current_successes >= self.n_successes
-        reward += self.reward_offset
-        return observation, reward, done, info
-
-    def compute_reward(self, achieved_goal, desired_goal, info):
-        reward = self.env.compute_reward(achieved_goal, desired_goal, info)
-        return reward + self.reward_offset
-
-
-class ActionNoiseWrapper(gym.Wrapper):
-    """
-    Add gaussian noise to the action (without telling the agent),
-    to test the robustness of the control.
-
-    :param env: (gym.Env)
-    :param noise_std: (float) Standard deviation of the noise
-    """
-
-    def __init__(self, env, noise_std=0.1):
-        super().__init__(env)
-        self.noise_std = noise_std
-
-    def step(self, action):
-        noise = np.random.normal(np.zeros_like(action), np.ones_like(action) * self.noise_std)
-        noisy_action = action + noise
-        return self.env.step(noisy_action)
-
-
-# from https://docs.obspy.org
-def lowpass(data, freq, df, corners=4, zerophase=False):
-    """
-    Butterworth-Lowpass Filter.
-
-    Filter data removing data over certain frequency ``freq`` using ``corners``
-    corners.
-    The filter uses :func:`scipy.signal.iirfilter` (for design)
-    and :func:`scipy.signal.sosfilt` (for applying the filter).
-
-    :type data: numpy.ndarray
-    :param data: Data to filter.
-    :param freq: Filter corner frequency.
-    :param df: Sampling rate in Hz.
-    :param corners: Filter corners / order.
-    :param zerophase: If True, apply filter once forwards and once backwards.
-        This results in twice the number of corners but zero phase shift in
-        the resulting filtered trace.
-    :return: Filtered data.
-    """
-    fe = 0.5 * df
-    f = freq / fe
-    # raise for some bad scenarios
-    if f > 1:
-        f = 1.0
-        msg = "Selected corner frequency is above Nyquist. " + "Setting Nyquist as high corner."
-        print(msg)
-    z, p, k = iirfilter(corners, f, btype="lowpass", ftype="butter", output="zpk")
-    sos = zpk2sos(z, p, k)
-    if zerophase:
-        firstpass = sosfilt(sos, data)
-        return sosfilt(sos, firstpass[::-1])[::-1]
-    else:
-        return sosfilt(sos, data)
-
-
-class LowPassFilterWrapper(gym.Wrapper):
-    """
-    Butterworth-Lowpass
-
-    :param env: (gym.Env)
-    :param freq: Filter corner frequency.
-    :param df: Sampling rate in Hz.
-    """
-
-    def __init__(self, env, freq=5.0, df=25.0):
-        super().__init__(env)
-        self.freq = freq
-        self.df = df
-        self.signal = []
-
-    def reset(self):
-        self.signal = []
-        return self.env.reset()
-
-    def step(self, action):
-        self.signal.append(action)
-        filtered = np.zeros_like(action)
-        for i in range(self.action_space.shape[0]):
-            smoothed_action = lowpass(np.array(self.signal)[:, i], freq=self.freq, df=self.df)
-            filtered[i] = smoothed_action[-1]
-        return self.env.step(filtered)
-
-
-class ActionSmoothingWrapper(gym.Wrapper):
-    """
-    Smooth the action using exponential moving average.
-
-    :param env: (gym.Env)
-    :param smoothing_coef: (float) Smoothing coefficient (0 no smoothing, 1 very smooth)
-    """
-
-    def __init__(self, env, smoothing_coef: float = 0.0):
-        super().__init__(env)
-        self.smoothing_coef = smoothing_coef
-        self.smoothed_action = None
-        # from https://github.com/rail-berkeley/softlearning/issues/3
-        # for smoothing latent space
-        # self.alpha = self.smoothing_coef
-        # self.beta = np.sqrt(1 - self.alpha ** 2) / (1 - self.alpha)
-
-    def reset(self):
-        self.smoothed_action = None
-        return self.env.reset()
-
-    def step(self, action):
-        if self.smoothed_action is None:
-            self.smoothed_action = np.zeros_like(action)
-        self.smoothed_action = self.smoothing_coef * self.smoothed_action + (1 - self.smoothing_coef) * action
-        return self.env.step(self.smoothed_action)
-
-
-class DelayedRewardWrapper(gym.Wrapper):
-    """
-    Delay the reward by `delay` steps, it makes the task harder but more realistic.
-    The reward is accumulated during those steps.
-
-    :param env: (gym.Env)
-    :param delay: (int) Number of steps the reward should be delayed.
-    """
-
-    def __init__(self, env, delay=10):
-        super().__init__(env)
-        self.delay = delay
-        self.current_step = 0
-        self.accumulated_reward = 0.0
-
-    def reset(self):
-        self.current_step = 0
-        self.accumulated_reward = 0.0
-        return self.env.reset()
-
-    def step(self, action):
-        observation, reward, terminated, truncated, info = self.env.step(action)
-        done = terminated or truncated
-
-        self.accumulated_reward += reward
-        self.current_step += 1
-
-        if self.current_step % self.delay == 0 or done:
-            reward = self.accumulated_reward
-            self.accumulated_reward = 0.0
-        else:
-            reward = 0.0
-        return observation, reward, done, info
-
-
-class HistoryWrapper(gym.Wrapper):
-    """
-    Stack past observations and actions to give an history to the agent.
-
-    :param env: (gym.Env)
-    :param horizon: (int) Number of steps to keep in the history.
-    """
-
-    def __init__(self, env: gym.Env, horizon: int = 5):
-        assert isinstance(env.observation_space, gym.spaces.Box)
-
-        wrapped_obs_space = env.observation_space
-        wrapped_action_space = env.action_space
-
-        # TODO: double check, it seems wrong when we have different low and highs
-        low_obs = np.repeat(wrapped_obs_space.low, horizon, axis=-1)
-        high_obs = np.repeat(wrapped_obs_space.high, horizon, axis=-1)
-
-        low_action = np.repeat(wrapped_action_space.low, horizon, axis=-1)
-        high_action = np.repeat(wrapped_action_space.high, horizon, axis=-1)
-
-        low = np.concatenate((low_obs, low_action))
-        high = np.concatenate((high_obs, high_action))
-
-        # Overwrite the observation space
-        env.observation_space = gym.spaces.Box(low=low, high=high, dtype=wrapped_obs_space.dtype)
-
-        super().__init__(env)
-
-        self.horizon = horizon
-        self.low_action, self.high_action = low_action, high_action
-        self.low_obs, self.high_obs = low_obs, high_obs
-        self.low, self.high = low, high
-        self.obs_history = np.zeros(low_obs.shape, low_obs.dtype)
-        self.action_history = np.zeros(low_action.shape, low_action.dtype)
-
-    def _create_obs_from_history(self):
-        return np.concatenate((self.obs_history, self.action_history))
-
-    def reset(self):
-        # Flush the history
-        self.obs_history[...] = 0
-        self.action_history[...] = 0
-        obs = self.env.reset()
-        self.obs_history[..., -obs.shape[-1] :] = obs
-        return self._create_obs_from_history()
-
-    def step(self, action):
-        obs, reward, terminated, truncated, info = self.env.step(action)
-        done = terminated or truncated
-        last_ax_size = obs.shape[-1]
-
-        self.obs_history = np.roll(self.obs_history, shift=-last_ax_size, axis=-1)
-        self.obs_history[..., -obs.shape[-1] :] = obs
-
-        self.action_history = np.roll(self.action_history, shift=-action.shape[-1], axis=-1)
-        self.action_history[..., -action.shape[-1] :] = action
-        return self._create_obs_from_history(), reward, done, info
-
-
-class HistoryWrapperObsDict(gym.Wrapper):
-    """
-    History Wrapper for dict observation.
-
-    :param env: (gym.Env)
-    :param horizon: (int) Number of steps to keep in the history.
-    """
-
-    def __init__(self, env, horizon=5):
-        assert isinstance(env.observation_space.spaces["observation"], gym.spaces.Box)
-
-        wrapped_obs_space = env.observation_space.spaces["observation"]
-        wrapped_action_space = env.action_space
-
-        # TODO: double check, it seems wrong when we have different low and highs
-        low_obs = np.repeat(wrapped_obs_space.low, horizon, axis=-1)
-        high_obs = np.repeat(wrapped_obs_space.high, horizon, axis=-1)
-
-        low_action = np.repeat(wrapped_action_space.low, horizon, axis=-1)
-        high_action = np.repeat(wrapped_action_space.high, horizon, axis=-1)
-
-        low = np.concatenate((low_obs, low_action))
-        high = np.concatenate((high_obs, high_action))
-
-        # Overwrite the observation space
-        env.observation_space.spaces["observation"] = gym.spaces.Box(low=low, high=high, dtype=wrapped_obs_space.dtype)
-
-        super().__init__(env)
-
-        self.horizon = horizon
-        self.low_action, self.high_action = low_action, high_action
-        self.low_obs, self.high_obs = low_obs, high_obs
-        self.low, self.high = low, high
-        self.obs_history = np.zeros(low_obs.shape, low_obs.dtype)
-        self.action_history = np.zeros(low_action.shape, low_action.dtype)
-
-    def _create_obs_from_history(self):
-        return np.concatenate((self.obs_history, self.action_history))
-
-    def reset(self):
-        # Flush the history
-        self.obs_history[...] = 0
-        self.action_history[...] = 0
-        obs_dict = self.env.reset()
-        obs = obs_dict["observation"]
-        self.obs_history[..., -obs.shape[-1] :] = obs
-
-        obs_dict["observation"] = self._create_obs_from_history()
-
-        return obs_dict
-
-    def step(self, action):
-        obs_dict, reward, terminated, truncated, info = self.env.step(action)
-        done = terminated or truncated
-        obs = obs_dict["observation"]
-        last_ax_size = obs.shape[-1]
-
-        self.obs_history = np.roll(self.obs_history, shift=-last_ax_size, axis=-1)
-        self.obs_history[..., -obs.shape[-1] :] = obs
-
-        self.action_history = np.roll(self.action_history, shift=-action.shape[-1], axis=-1)
-        self.action_history[..., -action.shape[-1] :] = action
-
-        obs_dict["observation"] = self._create_obs_from_history()
-
-        return obs_dict, reward, done, info
